+++ src/main/java/org/test/shapecalculator/common/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package org.test.shapecalculator;
+package org.test.shapecalculator.common;
 
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
@@ -6,13 +6,13 @@ import lombok.Getter;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 /**
- * The class represents geometric shape calculator measurement status for
- * the requests to which the application throws an exception.
+ * The class represents API error response for the requests to which
+ * the application throws an exception.
  */
 @ResponseBody
 @AllArgsConstructor
 @Getter
-public final class CalculatorMeasurementStatus {
+public final class ApiErrorResponse {
 
     @NotNull
     private String message;
