+++ src/main/java/org/test/shapecalculator/common/Deserializer.java
@@ -1,11 +1,11 @@
-package org.test.shapecalculator.calculator;
+package org.test.shapecalculator.common;
 
 /**
  * The JSON deserializer type of given boundaries.
  *
  * @param <T> the type of the object to which JSON is to  be deserialized.
  */
-interface Deserializer<T> {
+public interface Deserializer<T> {
 
     /**
      * Deserializes a given JSON into an object of the specified class.
