+++ src/main/java/org/test/shapecalculator/common/package-info.java
@@ -1,0 +1,4 @@
+@CheckReturnValue
+package org.test.shapecalculator.common;
+
+import com.google.errorprone.annotations.CheckReturnValue;
